+++ internal/workflow/workflow_test.go
@@ -4,7 +4,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/cschleiden/go-workflows/internal/converter"
+	"github.com/cschleiden/go-workflows/converter"
+	iconverter "github.com/cschleiden/go-workflows/internal/converter"
 	"github.com/cschleiden/go-workflows/internal/sync"
 	"github.com/cschleiden/go-workflows/internal/workflowerrors"
 	"github.com/stretchr/testify/require"
@@ -22,7 +23,7 @@ func Test_Workflow_WrapsPanic(t *testing.T) {
 	}
 
 	ctx := sync.Background()
-	ctx = converter.WithConverter(ctx, converter.DefaultConverter)
+	ctx = iconverter.WithConverter(ctx, converter.DefaultConverter)
 
 	wf := NewWorkflow(reflect.ValueOf(w))
 	err := wf.Execute(ctx, nil)
